Renamed actions to services
